Move CompilerBridgeProvider.constant into ZincCompilerUtil
On both my local machine and on Drone box, `ZincComponentCompilerSpec` fails to compile 2.10 compiler bridge with the following error:

```
/T/sbt_1234/xsbti/compile/CompilerBridgeProvider.java:33: error: `;' expected but `{' found.
    static CompilerBridgeProvider constant(File file, ScalaInstance scalaInstance) {
                                                                                   ^
```

I suspected that this is due to Java `interface` defining a static method, a change introduced in only in Java 8.
After making this change, the spec passed the test.
