+++ zinc/src/test/scala/sbt/inc/BaseCompilerSpec.scala
@@ -88,7 +88,7 @@ class BaseCompilerSpec extends BridgeProviderSpecification {
   }
 
   def scalaCompiler(instance: xsbti.compile.ScalaInstance, bridgeJar: File): AnalyzingCompiler = {
-    val bridgeProvider = CompilerBridgeProvider.constant(bridgeJar, instance)
+    val bridgeProvider = ZincUtil.constantBridgeProvider(instance, bridgeJar)
     val classpath = ClasspathOptionsUtil.boot
     val cache = Some(new ClassLoaderCache(new URLClassLoader(Array())))
     new AnalyzingCompiler(instance, bridgeProvider, classpath, _ => (), cache)
