+++ zinc/src/test/scala/sbt/inc/MultiProjectIncrementalSpec.scala
@@ -207,7 +207,7 @@ class MultiProjectIncrementalSpec extends BridgeProviderSpecification {
   }
 
   def scalaCompiler(instance: ScalaInstance, bridgeJar: File): AnalyzingCompiler = {
-    val bridgeProvider = CompilerBridgeProvider.constant(bridgeJar, instance)
+    val bridgeProvider = ZincUtil.constantBridgeProvider(instance, bridgeJar)
     val classpath = ClasspathOptionsUtil.boot
     val cache = Some(new ClassLoaderCache(new URLClassLoader(Array())))
     new AnalyzingCompiler(instance, bridgeProvider, classpath, _ => (), cache)
